Update test_json.py
The code here is for another lambda function that allowed us to send a pseudo file path to the back end.  The version here basically getting the what we want first before editing it in the actual lambda function for testing. It uses test.json as the basic test since we need to pull from files that looked very similar to that one when we actually run it.